[fix] #28 API 관련 Dto & Entity 수정

+++ src/main/java/com/kusher/kusher_in_korea/tabling/dto/request/UpdateRestaurantMenuDto.java
@@ -7,15 +7,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UpdateRestaurantMenuDto { // 식당 메뉴 수정 요청
     private Long menuId; // 메뉴번호
-
     private Long ownerId; // 유저번호 (식당 주인임을 확인하기 위함)
-
     private Long restaurantId; // 식당번호
-
     private String menuName; // 메뉴이름
-
-    private Long Price; // 메뉴가격
-
+    private int Price; // 메뉴가격
     private String Description; // 메뉴설명
-
 }
